REFACTOR: core/pkg/httputils -> core/pkg/httputil
