+++ core/internal/app/access/access_api.go
@@ -1,9 +1,9 @@
 package access
 
 import (
+	"core/internal/pkg/httputil"
 	rsc "core/internal/pkg/resource"
 	srv "core/internal/pkg/server"
-	"core/pkg/httputils"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -12,7 +12,7 @@ import (
 // ApplyRoutes access route handlers
 func ApplyRoutes(sctx *srv.Ctx, r *gin.RouterGroup) {
 	routes := r.Group(rsc.RouteAccess)
-	routes.POST("/token", httputils.Handler(sctx, generateTokenAPIHandler))
+	routes.POST("/token", httputil.Handler(sctx, generateTokenAPIHandler))
 }
 
 func generateTokenAPIHandler(sctx *srv.Ctx, ctx *gin.Context) {
