+++ core/internal/app/identity/identity_api.go
@@ -2,8 +2,8 @@ package identity
 
 import (
 	cons "core/internal/pkg/constants"
+	"core/internal/pkg/httputil"
 	rsc "core/internal/pkg/resource"
-	"core/pkg/httputils"
 	res "core/pkg/response"
 	"net/http"
 
@@ -13,12 +13,12 @@ import (
 // ApplyRoutes identity route handlers
 func ApplyRoutes(r *gin.RouterGroup) {
 	routes := r.Group(rsc.RouteIdentity)
-	rootPath := httputils.BuildPath(
+	rootPath := httputil.BuildPath(
 		rsc.WorkspaceKey,
 		rsc.ProjectKey,
 		rsc.EnvironmentKey,
 	)
-	resourcePath := httputils.AppendPath(
+	resourcePath := httputil.AppendPath(
 		rootPath,
 		rsc.IdentityKey,
 	)
@@ -31,22 +31,22 @@ func ApplyRoutes(r *gin.RouterGroup) {
 func listAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	data, _err := List(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
-		httputils.GetParam(ctx, rsc.EnvironmentKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.EnvironmentKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusCreated,
 		data,
@@ -58,23 +58,23 @@ func listAPIHandler(ctx *gin.Context) {
 func getAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	data, _err := Get(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
-		httputils.GetParam(ctx, rsc.EnvironmentKey),
-		httputils.GetParam(ctx, rsc.IdentityKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.EnvironmentKey),
+		httputil.GetParam(ctx, rsc.IdentityKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -86,22 +86,22 @@ func getAPIHandler(ctx *gin.Context) {
 func deleteAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	if err := Delete(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
-		httputils.GetParam(ctx, rsc.EnvironmentKey),
-		httputils.GetParam(ctx, rsc.IdentityKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.EnvironmentKey),
+		httputil.GetParam(ctx, rsc.IdentityKey),
 	); !err.IsEmpty() {
 		e.Extend(err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusNoContent,
 		&res.Success{},
