+++ core/internal/app/project/project_api.go
@@ -2,8 +2,8 @@ package project
 
 import (
 	cons "core/internal/pkg/constants"
+	"core/internal/pkg/httputil"
 	rsc "core/internal/pkg/resource"
-	"core/pkg/httputils"
 	"core/pkg/patch"
 	res "core/pkg/response"
 	"net/http"
@@ -14,10 +14,10 @@ import (
 // ApplyRoutes project route handlers
 func ApplyRoutes(r *gin.RouterGroup) {
 	routes := r.Group(rsc.RouteProject)
-	rootPath := httputils.BuildPath(
+	rootPath := httputil.BuildPath(
 		rsc.WorkspaceKey,
 	)
-	resourcePath := httputils.AppendPath(
+	resourcePath := httputil.AppendPath(
 		rootPath,
 		rsc.ProjectKey,
 	)
@@ -32,20 +32,20 @@ func ApplyRoutes(r *gin.RouterGroup) {
 func listAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	data, _err := List(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -57,7 +57,7 @@ func listAPIHandler(ctx *gin.Context) {
 func createAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
@@ -70,13 +70,13 @@ func createAPIHandler(ctx *gin.Context) {
 	data, _err := Create(
 		atk,
 		i,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusCreated,
 		data,
@@ -88,21 +88,21 @@ func createAPIHandler(ctx *gin.Context) {
 func getAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	data, _err := Get(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -115,7 +115,7 @@ func updateAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 	var i patch.Patch
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
@@ -127,14 +127,14 @@ func updateAPIHandler(ctx *gin.Context) {
 	data, _err := Update(
 		atk,
 		i,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -146,20 +146,20 @@ func updateAPIHandler(ctx *gin.Context) {
 func deleteAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	if err := Delete(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
-		httputils.GetParam(ctx, rsc.ProjectKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.ProjectKey),
 	); !err.IsEmpty() {
 		e.Extend(err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusNoContent,
 		&res.Success{},
