+++ core/internal/app/workspace/workspace_api.go
@@ -2,8 +2,8 @@ package workspace
 
 import (
 	cons "core/internal/pkg/constants"
+	"core/internal/pkg/httputil"
 	rsc "core/internal/pkg/resource"
-	"core/pkg/httputils"
 	"core/pkg/patch"
 	res "core/pkg/response"
 	"net/http"
@@ -14,7 +14,7 @@ import (
 // ApplyRoutes workspace route handlers
 func ApplyRoutes(r *gin.RouterGroup) {
 	routes := r.Group(rsc.RouteWorkspace)
-	resourcePath := httputils.BuildPath(
+	resourcePath := httputil.BuildPath(
 		rsc.WorkspaceKey,
 	)
 
@@ -28,7 +28,7 @@ func ApplyRoutes(r *gin.RouterGroup) {
 func listAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
@@ -38,7 +38,7 @@ func listAPIHandler(ctx *gin.Context) {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -50,7 +50,7 @@ func listAPIHandler(ctx *gin.Context) {
 func createAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
@@ -65,7 +65,7 @@ func createAPIHandler(ctx *gin.Context) {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusCreated,
 		data,
@@ -77,20 +77,20 @@ func createAPIHandler(ctx *gin.Context) {
 func getAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	data, _err := Get(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -103,7 +103,7 @@ func updateAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 	var i patch.Patch
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
@@ -115,13 +115,13 @@ func updateAPIHandler(ctx *gin.Context) {
 	data, _err := Update(
 		atk,
 		i,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
 	)
 	if !_err.IsEmpty() {
 		e.Extend(_err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusOK,
 		data,
@@ -133,19 +133,19 @@ func updateAPIHandler(ctx *gin.Context) {
 func deleteAPIHandler(ctx *gin.Context) {
 	var e res.Errors
 
-	atk, err := httputils.ExtractATK(ctx)
+	atk, err := httputil.ExtractATK(ctx)
 	if err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
 	if err := Delete(
 		atk,
-		httputils.GetParam(ctx, rsc.WorkspaceKey),
+		httputil.GetParam(ctx, rsc.WorkspaceKey),
 	); !err.IsEmpty() {
 		e.Extend(err)
 	}
 
-	httputils.Send(
+	httputil.Send(
 		ctx,
 		http.StatusNoContent,
 		&res.Success{},
