+++ core/internal/pkg/httputil/handler.go
@@ -1,4 +1,4 @@
-package httputils
+package httputil
 
 import (
 	srv "core/internal/pkg/server"
