+++ core/internal/pkg/httputil/params.go
@@ -1,4 +1,4 @@
-package httputils
+package httputil
 
 import (
 	rsc "core/internal/pkg/resource"
