+++ core/internal/pkg/httputil/send.go
@@ -1,4 +1,4 @@
-package httputils
+package httputil
 
 import (
 	res "core/pkg/response"
